+++ src/scenes/main_scene.lua
@@ -1,7 +1,10 @@
+local anim8         = require("lib.anim8")
 local moonshine     = require("lib.moonshine")
 local colors        = require("src.globals.colors")
 local consts        = require("src.globals.consts")
+local res           = require("src.globals.res")
 local drawer        = require("src.utils.drawer")
+local file          = require("src.utils.file")
 local input         = require("src.utils.input")
 
 local player        = require("src.models.player")
@@ -21,10 +24,17 @@ function mainScene:load(assets, actions, configs)
     self.isPaused                   = false
     self.isGameOver                 = false
 
+    -- === Shaders ===
     self.cmsShader                  = moonshine(moonshine.effects.chromasep)
     self.cmsShader.chromasep.radius = 4
 
-    player:init(self.assets.tileset)
+    -- === Animations ===
+    local tileW                     = self.assets.tileset:getWidth()
+    local tileH                     = self.assets.tileset:getHeight()
+    local grid                      = anim8.newGrid(16, 16, tileW, tileH)
+    local playerAnim                = anim8.newAnimation(grid("1-4", 1), 0.1)
+    self.enemyAnim                  = anim8.newAnimation(grid("1-3", 2), 0.2)
+    player:init(playerAnim)
 end
 
 function mainScene:reload()
@@ -50,9 +60,13 @@ function mainScene:handleInputs()
         self:reload()
     end
 
-    if input:isDown("left") or input:isDown("right") then
+    if input:wasPressed("right") and player.lane == 1 or
+        input:wasPressed("left") and player.lane == 2
+    then
         player:jumpStart()
-    elseif input:keyreleased("left") or input:keyreleased("right") then
+    end
+
+    if input:wasReleased("right") or input:wasReleased("left") then
         player:jumpEnd()
     end
 end
@@ -81,7 +95,7 @@ function mainScene:update(dt)
     self.obsTimer = self.obsTimer + dt
     if self.obsTimer > OBS_THRESHOLD then
         local m = tonumber(self.configs.mode)
-        local newObs = obstacle.get(m)
+        local newObs = obstacle.get(m, self.enemyAnim)
         table.insert(self.obstacles, newObs)
         self.obsTimer = 0
     end
@@ -100,32 +114,52 @@ function mainScene:draw()
     self.cmsShader(function()
         love.graphics.clear(colors.SLATE_100)
 
-        -- Draw lanes
+        -- === Draw lanes ===
         love.graphics.setColor(colors.SLATE_200)
         love.graphics.rectangle("fill", 0, 0, consts.LANE_WIDTH, consts.WINDOW_HEIGHT)
         local laneRightX = consts.WINDOW_WIDTH - consts.LANE_WIDTH
         love.graphics.rectangle("fill", laneRightX, 0, consts.LANE_WIDTH, consts.WINDOW_HEIGHT)
 
-        -- Draw center line
-        local cX = love.graphics.getWidth() / 2
+        -- === Draw center line ===
+        local centerX = love.graphics.getWidth() / 2
         love.graphics.setColor(colors.SLATE_200)
-        love.graphics.rectangle("line", cX, 0, 1, consts.WINDOW_HEIGHT)
+        love.graphics.rectangle("line", centerX, 0, 1, consts.WINDOW_HEIGHT)
 
+        -- === Draw player ===
         player:draw(self.assets.tileset)
 
-        -- Draw obstacles
+        -- === Draw obstacles ===
         for _, o in ipairs(self.obstacles) do
-            o:draw()
+            o:draw(self.assets.tileset)
         end
 
-        -- Draw pause screen
+        -- === Draw score ===
+        local font = file:getFont(res.MAIN_FONT, consts.FONT_HEADER_SIZE)
+        local scoreY = -280
+        local scoreW, scoreH = 150, 80
+        local centerY = love.graphics.getHeight() / 2
+
+        -- Draw score background
+        love.graphics.setColor(colors.SLATE_100)
+        love.graphics.rectangle(
+            "fill",
+            centerX - scoreH,
+            centerY + scoreY - scoreH / 2,
+            scoreW, scoreH
+        )
+
+        -- Draw score text
+        love.graphics.setColor(colors.SLATE_200)
+        drawer.drawCenteredText("99", font, 0, scoreY)
+
+        -- === Draw pause screen ===
         if self.isPaused then
             drawer.drawOverlay(140, "PAUSED", {
                 { text = "PRESS START", y = 0 },
             })
         end
 
-        -- Draw game over screen
+        -- === Draw game over screen ===
         if self.isGameOver then
             drawer.drawOverlay(140, "GAME OVER", {
                 { text = "PRESS START", y = 0 },