+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Revamped jumping logics, add visual guide, and catching some bugs 🐛
 - Update jumping control to feel more "natural"
 - Add animated scene transition
 - Add pseudo random obstacle generator
+- Add enemy animation
 - Minor bugs catched
 
 ## 0.0.3